# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Platform API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Business {
  id: number;
  userId: string;
  businessName: string;
  ownerName: string;
  subscriberName: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  description?: string | null;
  onboardingComplete: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface BusinessInput {
  businessName: string;
  ownerName: string;
  subscriberName: string;
  industry?: string;
  description?: string;
}

export interface BusinessUpdate {
  businessName?: string;
  ownerName?: string;
  subscriberName?: string;
  industry?: string;
  description?: string;
  onboardingComplete?: boolean;
}

export interface GbpProfile {
  id: number;
  businessId: number;
  /** @nullable */
  placeId?: string | null;
  businessName: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  category?: string | null;
  isVerified: boolean;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  website?: string | null;
  createdAt: string;
}

export interface GbpProfileInput {
  businessName: string;
  placeId?: string;
  address?: string;
  category?: string;
  isVerified: boolean;
  phoneNumber?: string;
  website?: string;
}

export interface GbpProfileUpdate {
  businessName?: string;
  address?: string;
  category?: string;
  isVerified?: boolean;
  phoneNumber?: string;
  website?: string;
}

export type WebsiteLinkType =
  (typeof WebsiteLinkType)[keyof typeof WebsiteLinkType];

export const WebsiteLinkType = {
  backlink: "backlink",
  blog: "blog",
  newsletter: "newsletter",
  news: "news",
  social: "social",
  directory: "directory",
  other: "other",
} as const;

export interface Website {
  id: number;
  businessId: number;
  url: string;
  linkType: WebsiteLinkType;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  domainAuthority?: number | null;
  createdAt: string;
}

export type WebsiteInputLinkType =
  (typeof WebsiteInputLinkType)[keyof typeof WebsiteInputLinkType];

export const WebsiteInputLinkType = {
  backlink: "backlink",
  blog: "blog",
  newsletter: "newsletter",
  news: "news",
  social: "social",
  directory: "directory",
  other: "other",
} as const;

export interface WebsiteInput {
  url: string;
  linkType: WebsiteInputLinkType;
  title?: string;
  domainAuthority?: number;
}

export type WebsiteUpdateLinkType =
  (typeof WebsiteUpdateLinkType)[keyof typeof WebsiteUpdateLinkType];

export const WebsiteUpdateLinkType = {
  backlink: "backlink",
  blog: "blog",
  newsletter: "newsletter",
  news: "news",
  social: "social",
  directory: "directory",
  other: "other",
} as const;

export interface WebsiteUpdate {
  url?: string;
  linkType?: WebsiteUpdateLinkType;
  title?: string;
  domainAuthority?: number;
}

export type KeywordStatus = (typeof KeywordStatus)[keyof typeof KeywordStatus];

export const KeywordStatus = {
  active: "active",
  paused: "paused",
  archived: "archived",
} as const;

export interface Keyword {
  id: number;
  businessId: number;
  keyword: string;
  /** @nullable */
  efficiencyScore?: number | null;
  /** @nullable */
  searchVolume?: number | null;
  /** @nullable */
  currentPosition?: number | null;
  /** @nullable */
  previousPosition?: number | null;
  isAiGenerated: boolean;
  /** @nullable */
  notes?: string | null;
  status: KeywordStatus;
  createdAt: string;
  updatedAt: string;
}

export interface KeywordInput {
  keyword: string;
  notes?: string;
  isAiGenerated?: boolean;
}

export type KeywordUpdateStatus =
  (typeof KeywordUpdateStatus)[keyof typeof KeywordUpdateStatus];

export const KeywordUpdateStatus = {
  active: "active",
  paused: "paused",
  archived: "archived",
} as const;

export interface KeywordUpdate {
  keyword?: string;
  notes?: string;
  status?: KeywordUpdateStatus;
}

export interface KeywordSuggestion {
  keyword: string;
  reason: string;
  /** @nullable */
  estimatedVolume?: number | null;
  efficiencyScore: number;
}

export interface ReportKeyword {
  keyword: string;
  /** @nullable */
  position?: number | null;
  /** @nullable */
  change?: number | null;
  /** @nullable */
  impressions?: number | null;
}

export interface Report {
  id: number;
  businessId: number;
  periodStart: string;
  periodEnd: string;
  /** @nullable */
  visibilityScore?: number | null;
  /** @nullable */
  totalImpressions?: number | null;
  /** @nullable */
  totalClicks?: number | null;
  /** @nullable */
  averagePosition?: number | null;
  keywordsTracked: number;
  keywordsImproved: number;
  keywordsDeclined: number;
  /** @nullable */
  aiSummary?: string | null;
  topKeywords: ReportKeyword[];
  createdAt: string;
}

export type DashboardSummaryRecentKeywordTrend =
  (typeof DashboardSummaryRecentKeywordTrend)[keyof typeof DashboardSummaryRecentKeywordTrend];

export const DashboardSummaryRecentKeywordTrend = {
  improving: "improving",
  declining: "declining",
  stable: "stable",
  unknown: "unknown",
} as const;

export interface DashboardSummary {
  totalKeywords: number;
  activeKeywords: number;
  /** @nullable */
  averageEfficiencyScore?: number | null;
  /** @nullable */
  visibilityScore?: number | null;
  /** @nullable */
  visibilityChange?: number | null;
  totalWebsites: number;
  gbpVerified: boolean;
  recentKeywordTrend: DashboardSummaryRecentKeywordTrend;
  /** @nullable */
  lastReportDate?: string | null;
  topKeywords?: Keyword[];
  onboardingComplete: boolean;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}