# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEO Platform API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/me:
    get:
      operationId: getMyBusiness
      tags: [business]
      summary: Get the current user's business profile
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createBusiness
      tags: [business]
      summary: Create a business profile (onboarding step 1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      operationId: updateBusiness
      tags: [business]
      summary: Update the business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"

  # ─── GBP ─────────────────────────────────────────────────────────────────────
  /businesses/me/gbp:
    get:
      operationId: listGbpProfiles
      tags: [business]
      summary: List GBP profiles for the current business
      responses:
        "200":
          description: GBP profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GbpProfile"
    post:
      operationId: addGbpProfile
      tags: [business]
      summary: Add a Google Business Profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GbpProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GbpProfile"
  /businesses/me/gbp/{id}:
    patch:
      operationId: updateGbpProfile
      tags: [business]
      summary: Update a GBP profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GbpProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GbpProfile"
    delete:
      operationId: deleteGbpProfile
      tags: [business]
      summary: Remove a GBP profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Websites ─────────────────────────────────────────────────────────────────
  /businesses/me/websites:
    get:
      operationId: listWebsites
      tags: [business]
      summary: List websites for the current business
      responses:
        "200":
          description: Websites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Website"
    post:
      operationId: addWebsite
      tags: [business]
      summary: Add a website/link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
  /businesses/me/websites/{id}:
    patch:
      operationId: updateWebsite
      tags: [business]
      summary: Update a website entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
    delete:
      operationId: deleteWebsite
      tags: [business]
      summary: Remove a website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Keywords ─────────────────────────────────────────────────────────────────
  /businesses/me/keywords:
    get:
      operationId: listKeywords
      tags: [keywords]
      summary: List keywords for the current business
      responses:
        "200":
          description: Keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: addKeyword
      tags: [keywords]
      summary: Add a keyword manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
...
  /businesses/me/keywords/{id}:
    patch:
      operationId: updateKeyword
      tags: [keywords]
      summary: Update a keyword (edit text, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
    delete:
      operationId: deleteKeyword
      tags: [keywords]
      summary: Remove a keyword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Reports ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Business ───────────────────────────────────────────────────────────────
    Business:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        businessName:
          type: string
        ownerName:
          type: string
        subscriberName:
          type: string
        industry:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - businessName
        - ownerName
        - subscriberName
        - onboardingComplete
        - createdAt
        - updatedAt
    BusinessInput:
      type: object
      properties:
        businessName:
          type: string
        ownerName:
          type: string
        subscriberName:
          type: string
        industry:
          type: string
        description:
          type: string
      required:
        - businessName
        - ownerName
        - subscriberName
    BusinessUpdate:
      type: object
      properties:
        businessName:
          type: string
        ownerName:
          type: string
        subscriberName:
          type: string
        industry:
          type: string
        description:
          type: string
        onboardingComplete:
          type: boolean

    # ─── GBP ────────────────────────────────────────────────────────────────────
    GbpProfile:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        placeId:
          type: ["string", "null"]
        businessName:
          type: string
        address:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        isVerified:
          type: boolean
        phoneNumber:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - businessName
        - isVerified
        - createdAt
    GbpProfileInput:
      type: object
      properties:
        businessName:
          type: string
        placeId:
          type: string
        address:
          type: string
        category:
          type: string
        isVerified:
          type: boolean
        phoneNumber:
          type: string
        website:
          type: string
      required:
        - businessName
        - isVerified
    GbpProfileUpdate:
      type: object
      properties:
        businessName:
          type: string
        address:
          type: string
        category:
          type: string
        isVerified:
          type: boolean
        phoneNumber:
          type: string
        website:
          type: string

    # ─── Website ─────────────────────────────────────────────────────────────────
    Website:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        url:
          type: string
        linkType:
          type: string
          enum: [backlink, blog, newsletter, news, social, directory, other]
        title:
          type: ["string", "null"]
        domainAuthority:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - url
        - linkType
        - createdAt
    WebsiteInput:
      type: object
      properties:
        url:
          type: string
        linkType:
          type: string
          enum: [backlink, blog, newsletter, news, social, directory, other]
        title:
          type: string
        domainAuthority:
          type: integer
      required:
        - url
        - linkType
    WebsiteUpdate:
      type: object
      properties:
        url:
          type: string
        linkType:
          type: string
          enum: [backlink, blog, newsletter, news, social, directory, other]
        title:
          type: string
        domainAuthority:
          type: integer

    # ─── Keywords ────────────────────────────────────────────────────────────────
    Keyword:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        keyword:
          type: string
        efficiencyScore:
          type: ["number", "null"]
        searchVolume:
          type: ["integer", "null"]
        currentPosition:
          type: ["integer", "null"]
        previousPosition:
          type: ["integer", "null"]
        isAiGenerated:
          type: boolean
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - keyword
        - isAiGenerated
        - status
        - createdAt
        - updatedAt
    KeywordInput:
      type: object
      properties:
        keyword:
          type: string
        notes:
          type: string
        isAiGenerated:
          type: boolean
      required:
        - keyword
    KeywordUpdate:
      type: object
      properties:
        keyword:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [active, paused, archived]
...