# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Platform API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Business {
  id: number;
  userId: string;
  businessName: string;
  ownerName: string;
  subscriberName: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  description?: string | null;
  onboardingComplete: boolean;
  createdAt: Date;
  updatedAt: Date;
}