# lib/api-zod/src/generated/types/businessInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Platform API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessInput {
  businessName: string;
  ownerName: string;
  subscriberName: string;
  industry?: string;
  description?: string;
}