# lib/api-zod/src/generated/types/gbpProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Platform API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GbpProfileInput {
  businessName: string;
  placeId?: string;
  address?: string;
  category?: string;
  isVerified: boolean;
  phoneNumber?: string;
  website?: string;
}