# lib/api-zod/src/generated/types/gbpProfileUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Platform API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GbpProfileUpdate {
  businessName?: string;
  address?: string;
  category?: string;
  isVerified?: boolean;
  phoneNumber?: string;
  website?: string;
}