# lib/api-zod/src/generated/types/keywordInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Platform API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KeywordInput {
  keyword: string;
  notes?: string;
  isAiGenerated?: boolean;
}