# lib/api-zod/src/generated/types/keywordUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Platform API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KeywordUpdateStatus } from "./keywordUpdateStatus";

export interface KeywordUpdate {
  keyword?: string;
  notes?: string;
  status?: KeywordUpdateStatus;
}