# lib/api-zod/src/generated/types/websiteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Platform API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WebsiteInputLinkType } from "./websiteInputLinkType";

export interface WebsiteInput {
  url: string;
  linkType: WebsiteInputLinkType;
  title?: string;
  domainAuthority?: number;
}